+++ src/main/resources/test_examples/exampleGrammar.json
@@ -1,7 +1,8 @@
 {
   "terminals": [
     "(",
-    ")"
+    ")",
+    "epsilon"
   ],
   "nonTerminals": [
     "S"
